+++ src/App.css
@@ -1,12 +1,3 @@
-/* .gradient {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    height: 100%;
-  } */
-
-
 @import url('https://fonts.googleapis.com/css2?family=Kufam:ital,wght@0,400..900;1,400..900&display=swap');
 
   .gradient {
