+++ src/App.tsx
@@ -1,16 +1,15 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import Home from "./routes/Home";
-import Temp from "./routes/Template";
+import Hype from "./routes/hype";
 
 function App() {
 	return (
 		<BrowserRouter>
 			<Routes>
-				<Route path="/auth/" element={<Temp text="auth"/>}> </Route>
-				<Route path="/schedule/" element={<Temp text="schedule"/>}> </Route>
-				<Route path="/questions/" element={<Temp text="questions"/>}> </Route>
-				<Route path="/" element={<Home/>} />
-				<Route path="*" element={<Temp text="404" />} />
+				<Route path="/auth/" element={<Hype/>}> </Route>
+				<Route path="/schedule/" element={<Hype/>}> </Route>
+				<Route path="/questions/" element={<Hype/>}> </Route>
+				<Route path="/" element={<Hype/>} />
+				<Route path="*" element={<Hype/>} />
 			</Routes>
 		</BrowserRouter>
 	);
